refactor: rolepermission model
