+++ interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from "./category"
 export * from "./supplier"
 export * from "./user"
+export * from "./rolePermission"
